+++ mollie.php
@@ -26,17 +26,16 @@ use Mollie\Repository\MolOrderPaymentFeeRepositoryInterface;
 use Mollie\Repository\PaymentMethodRepositoryInterface;
 use Mollie\Service\ExceptionService;
 use Mollie\ServiceProvider\LeagueServiceContainerProvider;
-use Mollie\Subscription\Exception\SubscriptionProductValidationException;
 use Mollie\Subscription\Handler\CustomerAddressUpdateHandler;
+use Mollie\Subscription\Handler\UpdateSubscriptionCarrierHandler;
 use Mollie\Subscription\Install\AttributeInstaller;
 use Mollie\Subscription\Install\DatabaseTableInstaller;
 use Mollie\Subscription\Install\HookInstaller;
 use Mollie\Subscription\Install\Installer;
-use Mollie\Subscription\Logger\NullLogger;
+use Mollie\Subscription\Provider\SubscriptionProductProvider;
 use Mollie\Subscription\Repository\LanguageRepository as LanguageAdapter;
 use Mollie\Subscription\Repository\RecurringOrderRepositoryInterface;
 use Mollie\Subscription\Validator\CanProductBeAddedToCartValidator;
-use Mollie\Subscription\Verification\HasSubscriptionProductInCart;
 use Mollie\Utility\PsVersionUtility;
 use Mollie\Verification\IsPaymentInformationAvailable;
 use PrestaShop\PrestaShop\Core\Localization\Locale\Repository;
@@ -85,7 +84,7 @@ class Mollie extends PaymentModule
     {
         $this->name = 'mollie';
         $this->tab = 'payments_gateways';
-        $this->version = '6.1.1';
+        $this->version = '6.2.0';
         $this->author = 'Mollie B.V.';
         $this->need_instance = 1;
         $this->bootstrap = true;
@@ -180,7 +179,7 @@ class Mollie extends PaymentModule
         $subscriptionInstaller = new Installer(
             new DatabaseTableInstaller(),
             new AttributeInstaller(
-                new NullLogger(),
+                $this->getService(PrestaLoggerInterface::class),
                 $this->getService(ConfigurationAdapter::class),
                 $this,
                 new LanguageAdapter(),
@@ -1014,14 +1013,14 @@ class Mollie extends PaymentModule
         return '';
     }
 
-    public function hookActionCartUpdateQuantityBefore($params)
+    public function hookActionCartUpdateQuantityBefore($params): void
     {
-        /** @var CanProductBeAddedToCartValidator $cartValidation */
-        $cartValidation = $this->getService(CanProductBeAddedToCartValidator::class);
+        /** @var CanProductBeAddedToCartValidator $canProductBeAddedToCartValidator */
+        $canProductBeAddedToCartValidator = $this->getService(CanProductBeAddedToCartValidator::class);
 
         try {
-            $cartValidation->validate((int) $params['id_product_attribute']);
-        } catch (SubscriptionProductValidationException $e) {
+            $canProductBeAddedToCartValidator->validate((int) $params['id_product_attribute']);
+        } catch (\Throwable $exception) {
             $product = $this->makeProductNotOrderable($params['product']);
 
             $params['product'] = $product;
@@ -1273,6 +1272,39 @@ class Mollie extends PaymentModule
         $this->addPreventDeleteErrorMessage();
     }
 
+    public function hookActionCarrierUpdate(array $params): void
+    {
+        $oldCarrierId = $params['id_carrier'] ?? 0;
+        $newCarrier = $params['carrier'] ?? null;
+
+        if (empty($oldCarrierId) || empty($newCarrier)) {
+            return;
+        }
+
+        /** @var UpdateSubscriptionCarrierHandler $subscriptionCarrierUpdateHandler */
+        $subscriptionCarrierUpdateHandler = $this->getService(UpdateSubscriptionCarrierHandler::class);
+
+        /** @var ConfigurationAdapter $configuration */
+        $configuration = $this->getService(ConfigurationAdapter::class);
+
+        /** @var PrestaLoggerInterface $logger */
+        $logger = $this->getService(PrestaLoggerInterface::class);
+
+        if ((int) $oldCarrierId !== (int) $configuration->get(Config::MOLLIE_SUBSCRIPTION_ORDER_CARRIER_ID)) {
+            return;
+        }
+
+        $failedSubscriptionOrderIdsToUpdate = $subscriptionCarrierUpdateHandler->run((int) $newCarrier->id);
+
+        if (empty($failedSubscriptionOrderIdsToUpdate)) {
+            return;
+        }
+
+        $logger->error('Failed to update subscription carrier for all orders.', [
+            'failed_subscription_order_ids' => json_encode($failedSubscriptionOrderIdsToUpdate),
+        ]);
+    }
+
     public function hookActionFrontControllerAfterInit(): void
     {
         $this->frontControllerAfterInit();
@@ -1293,15 +1325,19 @@ class Mollie extends PaymentModule
             return;
         }
 
-        /** @var HasSubscriptionProductInCart $hasSubscriptionProductInCart */
-        $hasSubscriptionProductInCart = $this->getService(HasSubscriptionProductInCart::class);
+        if (empty($this->context->cart) || empty($this->context->cart->getProducts())) {
+            return;
+        }
+
+        /** @var SubscriptionProductProvider $subscriptionProductProvider */
+        $subscriptionProductProvider = $this->getService(SubscriptionProductProvider::class);
+
+        if (empty($subscriptionProductProvider->getProduct($this->context->cart->getProducts()))) {
+            return;
+        }
 
         /** @var Link $link */
         $link = $this->getService(Link::class);
-
-        if (!$hasSubscriptionProductInCart->verify()) {
-            return;
-        }
 
         $this->context->controller->warning[] = $this->l('Customer must be logged in to buy subscription item.');