+++ src/Config/Config.php
@@ -110,6 +110,7 @@ class Config
         'sandbox' => 'MOLLIE_SANDBOX_SINGLE_CLICK_PAYMENT',
         'production' => 'MOLLIE_PRODUCTION_SINGLE_CLICK_PAYMENT',
     ];
+
     const MOLLIE_IMAGES = 'MOLLIE_IMAGES';
     const MOLLIE_SHOW_RESEND_PAYMENT_LINK = 'MOLLIE_SHOW_RESEND_PAYMENT_LINK';
     const MOLLIE_ISSUERS = [
@@ -162,6 +163,7 @@ class Config
     const MOLLIE_CARRIER_CUSTOM_URL = 'MOLLIE_CARRIER_CUSTOM_URL_';
 
     const MOLLIE_SUBSCRIPTION_ORDER_CARRIER_ID = 'MOLLIE_SUBSCRIPTION_ORDER_CARRIER_ID';
+    const MOLLIE_SUBSCRIPTION_ENABLED = 'MOLLIE_SUBSCRIPTION_ENABLED';
 
     const MOLLIE_METHOD_ENABLED = 'MOLLIE_METHOD_ENABLED_';
     const MOLLIE_METHOD_TITLE = 'MOLLIE_METHOD_TITLE_';
@@ -280,6 +282,10 @@ class Config
         self::MOLLIE_in3_METHOD_ID,
     ];
 
+    const PAYMENT_API_ONLY_METHODS = [
+        PaymentMethod::ALMA,
+    ];
+
     const ROUTE_RESEND_SECOND_CHANCE_PAYMENT_MESSAGE = 'mollie_module_admin_resend_payment_message';
 
     const PAYMENT_FEE_SKU = 'payment-fee-sku';
@@ -317,6 +323,8 @@ class Config
         'in3' => 'in3',
         'billie' => 'Billie',
         'twint' => 'TWINT',
+        'bancomat' => 'Bancomat',
+        'alma' => 'Alma',
         'blik' => 'BLIK',
         'klarna' => 'Pay with Klarna.',
     ];