+++ src/Service/SettingsSaveService.php
@@ -243,6 +243,7 @@ class SettingsSaveService
         $mollieOrderConfirmationSand = $this->tools->getValue(Config::MOLLIE_SEND_ORDER_CONFIRMATION);
         $mollieIFrameEnabled = $this->tools->getValue(Config::MOLLIE_IFRAME[$environment ? 'production' : 'sandbox']);
         $mollieSingleClickPaymentEnabled = $this->tools->getValue(Config::MOLLIE_SINGLE_CLICK_PAYMENT[$environment ? 'production' : 'sandbox']);
+
         $mollieImages = $this->tools->getValue(Config::MOLLIE_IMAGES);
         $showResentPayment = $this->tools->getValue(Config::MOLLIE_SHOW_RESEND_PAYMENT_LINK);
         $mollieIssuers = $this->tools->getValue(Config::MOLLIE_ISSUERS[$environment ? 'production' : 'sandbox']);
@@ -260,8 +261,6 @@ class SettingsSaveService
         $voucherCategory = $this->tools->getValue(Config::MOLLIE_VOUCHER_CATEGORY);
         $applePayDirectStyle = $this->tools->getValue(Config::MOLLIE_APPLE_PAY_DIRECT_STYLE);
         $isBancontactQrCodeEnabled = $this->tools->getValue(Config::MOLLIE_BANCONTACT_QR_CODE_ENABLED);
-
-        $subscriptionsShippingOption = (int) $this->tools->getValue(Config::MOLLIE_SUBSCRIPTION_ORDER_CARRIER_ID);
 
         $mollieShipMain = $this->tools->getValue(Config::MOLLIE_AUTO_SHIP_MAIN);
         if (!isset($mollieErrors)) {
@@ -317,7 +316,6 @@ class SettingsSaveService
             $this->configurationAdapter->updateValue(Config::MOLLIE_DEBUG_LOG, (int) $mollieLogger);
             $this->configurationAdapter->updateValue(Config::MOLLIE_API, $mollieApi);
             $this->configurationAdapter->updateValue(Config::MOLLIE_VOUCHER_CATEGORY, $voucherCategory);
-            $this->configurationAdapter->updateValue(Config::MOLLIE_SUBSCRIPTION_ORDER_CARRIER_ID, $subscriptionsShippingOption);
             $this->configurationAdapter->updateValue(
                 Config::MOLLIE_AUTO_SHIP_STATUSES,
                 json_encode($this->getStatusesValue(Config::MOLLIE_AUTO_SHIP_STATUSES))