+++ changelog.md
@@ -2,6 +2,16 @@
 
 # Changelog #
 
+## Changes in release 6.2.0 ##
++ New payment methods: Bancomat and Alma
++ Apple certificate update
++ Conflicting services fix
++ Bootstrap 5 compatibility improved
++ Maximum fee field
++ Multi carrier for subscription orders added
++ Design improvements in BO
++ Improved subscription creation logic
+
 ## Changes in release 6.1.1 ##
 + Updated translations for Dutch, German, English and French languages
 + Added credit card translations for Italian and Spanish languages